Moved test folders, moved root scripts into utils and merged deployment with server and CLI. Paper trade scripts are now inside of paper trade.
