+++ Blankly/strategy/cookbook/recipes.py
