--- Blankly/utils/ci_utils/Settings.json
@@ -1,19 +1,0 @@
-{
-  "settings": {
-    "account_update_time": 5000,
-    "use_sandbox": true,
-    "use_sandbox_websockets": false,
-    "websocket_buffer_size": 10000,
-
-    "coinbase_pro": {
-      "cash": "USD"
-    },
-    "binance": {
-      "cash": "USDT",
-      "binance_tld": "us"
-    },
-    "alpaca": {
-      "cash": "USD"
-    }
-  }
-}
