--- Blankly/utils/ci_utils/requirements.txt
@@ -1,8 +1,0 @@
-numpy
-sklearn
-scikit-learn
-zerorpc
-requests
-websocket-client
-pandas
-python-binance
