--- Blankly/utils/ci_utils/tests_controller.py
@@ -1,10 +1,0 @@
-import unittest
-from root.test_auth_constructor import TestAuthConstructor
-
-from root.exchanges.test_interface_homogeneity import InterfaceHomogeneity
-from root.exchanges.Coinbase_Pro.test_Coinbase_Pro_Interface import CoinbaseInterface
-from root.exchanges.Binance.test_Binance_Interface import BinanceInterface
-
-
-if __name__ == "__main__":
-    unittest.main()
